tests: fix the same shared-VFS-wipe hazard in test_vfs_fs_backed.py
Its autouse teardown called scene.albam.vfs.file_list.clear() and
fs_registry.clear() unconditionally - issue #271 already named this as
the same hazard fixed in tests/cie/ (see that commit): wiping state
shared with every other engine's session-scoped test fixtures, not just
this file's own.

Moved close_new_fs_roots/vfs_root_names/remove_new_vfs_roots from
tests/cie/conftest.py to the root tests/conftest.py, now that a second
consumer needs them, and switched this file's teardown to use them.

Also fixes test_remove_root_unregisters_fs, which asserted the whole VFS
was empty after removing its own root - true only by accident of the old
wholesale-clear teardown. Now checks its own root is gone and the list
size is back to what it was before, which is what the test actually
means to verify.

Direct in-test fs_registry.clear() calls elsewhere in this file are left
alone - those are deliberately testing recovery from a real registry
clear (Reload Scripts, process restart), not incidental cleanup.

Verified against the exact 5-app_id CI invocation, not just this file in
isolation: 664 passed, 0 failed.
